requirements.txt
opencv-python
ultralytics
transformers
torch
pyttsx3
Pillow
fastapi
uvicorn
python-multipart
pytesseract
